[Resolved] Serializable problem
